requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
bottleneck=1.3.5=py39h67323c0_0
brotli-python=1.0.9=py39he9d5cce_7
ca-certificates=2023.12.12=hecd8cb5_0
certifi=2023.11.17=py39hecd8cb5_0
cffi=1.16.0=py39h6c40b1e_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=41.0.7=py39h30e54ef_0
filelock=3.13.1=py39hecd8cb5_0
freetype=2.12.1=hd8bbffd_0
fsspec=2024.2.0=pypi_0
giflib=5.2.1=h6c40b1e_3
gmp=6.2.1=he9d5cce_3
gmpy2=2.1.2=py39hd5de756_0
idna=3.4=py39hecd8cb5_0
intel-openmp=2023.1.0=ha357a0b_43548
jinja2=3.1.3=py39hecd8cb5_0
jpeg=9e=h6c40b1e_1
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libcxx=14.0.6=h9765a3e_0
libdeflate=1.17=hb664fd8_1
libffi=3.4.4=hecd8cb5_0
libjpeg-turbo=2.0.0=hca72f7f_0
libpng=1.6.39=h6c40b1e_0
libtiff=4.5.1=hcec6c5f_0
libwebp=1.3.2=hf6ce154_0
libwebp-base=1.3.2=h6c40b1e_0
lightning-utilities=0.10.1=pypi_0
lz4-c=1.9.4=hcec6c5f_0
markupsafe=2.1.3=py39h6c40b1e_0
mkl=2023.1.0=h8e150cf_43560
mkl-service=2.4.0=py39h6c40b1e_1
mkl_fft=1.3.8=py39h6c40b1e_0
mkl_random=1.2.4=py39ha357a0b_0
mpc=1.1.0=h6ef4df4_1
mpfr=4.0.2=h9066e36_1
mpmath=1.3.0=py39hecd8cb5_0
ncurses=6.4=hcec6c5f_0
networkx=3.1=py39hecd8cb5_0
numexpr=2.8.7=py39h827a554_0
numpy=1.26.3=py39h827a554_0
numpy-base=1.26.3=py39ha186be2_0
openjpeg=2.4.0=h66ea3da_0
openssl=3.0.12=hca72f7f_0
packaging=23.2=pypi_0
pandas=2.1.4=py39h3ea8b11_0
pillow=10.0.1=py39h7d39338_0
pip=23.3.1=py39hecd8cb5_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py39hecd8cb5_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.18=h5ee71fb_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.2.0=py3.9_0
pytz=2023.3.post1=py39hecd8cb5_0
pyyaml=6.0.1=py39h6c40b1e_0
readline=8.2=hca72f7f_0
requests=2.31.0=py39hecd8cb5_0
setuptools=68.2.2=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h6c40b1e_0
sympy=1.12=py39hecd8cb5_0
tbb=2021.8.0=ha357a0b_0
tk=8.6.12=h5d9f67b_0
torcheval=0.0.7=pypi_0
torchmetrics=1.3.0.post0=pypi_0
torchvision=0.17.0=py39_cpu
typing_extensions=4.9.0=py39hecd8cb5_1
tzdata=2023d=h04d1e81_0
urllib3=1.26.18=py39hecd8cb5_0
wheel=0.41.2=py39hecd8cb5_0
xz=5.4.5=h6c40b1e_0
yaml=0.2.5=haf1e3a3_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.5=hc035e20_0
